+++ Sources/Configuration.swift
@@ -153,7 +153,7 @@ struct StateStore {
         try FileManager.default.setAttributes([.posixPermissions: 0o700], ofItemAtPath: directory.path)
         let encoder = JSONEncoder()
         encoder.outputFormatting = [.prettyPrinted, .sortedKeys]
-        try encoder.encode(state).write(to: file, options: [.atomic, .completeFileProtectionUnlessOpen])
+        try encoder.encode(state).write(to: file, options: .atomic)
         try FileManager.default.setAttributes([.posixPermissions: 0o600], ofItemAtPath: file.path)
     }
     func freshState() throws -> SavedState {